requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==1.24.3
opencv-python==4.8.1.78
ultralytics==8.0.232
roboflow==1.1.9
pydantic==2.5.0
Pillow==9.5.0